hikey960: Enabling dynamic partitions for HiKey960
This is basically a port of work Amit had done against HiKey
board awhile back:
  https://android-review.linaro.org/c/device/linaro/hikey/+/20550

Moving system vendor product and system_ext to the super
partition.

For now, instead of reworking the ptable, we simply flash the
super partition to the system partition. This will need to change
so that we change the ptable as well, but that needs some time
as the ptable flashing can be disruptive for testing.

Feedback would be appreciated.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: Ie53fed6392116b16063f497978e369490a1fccd2