hikey960: fstab.hikey960 Add new dynamic partitions to the vendor fstab
While we boot up properly with the dynamic partitions in only
the ramdisk fstab, Yongqin noted adb remount was failing recently.

On looking into the issue, I realized one issue is we don't have
the dynamic paritions listed in the vendor fstab which causes
some of the remount logic to not find the partitions.

This patch adds the dynamic partitions and unifies the fstab
file for hikey960 so we have the same one for both the ramdisk
and vendor image.

NOTE: adb remount still does not fully work with this change. We
will need a follow on patch to fix that.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: Idf9c463162398ef89db31fbf69d205d1a47e2cdd