hikey960: fstab.hikey960: Set metadata as first_stage_mount to resolve adb remount overlay persistence issues
It was reported that changes made after "adb remount" were not
persisting after a reboot. After reproducing and debugging, I
was at a bit of a loss for why.

After quite a bit of floundering around, I finally realized we
were missing first_stage_mount from the fstab arguments. This
prevents the metadata partition from being available early
enough for the overlay images to be located and applied at
bootup.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I63a8cedacbace33cceee7f83121a4b83dd996cc3