hikey960: fstab.hikey960: Drop f2fs userdata
Over the last week I started seeing some occasional trouble
with the f2fs entry in fstab.hikey960 causing init confusion
where the fsck found bad magic value (since the partition is
ext4), but the partition wasn't released before the ext4
mount began and saw it as in use.

This caused userdata not not be mounted at all, and resulted
in a number of keystore errors and the system not booting
fully to UI.

I don't think anyone is really using the f2fs userdata config
with HiKey960, so lets drop it.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: Ibc34b74a252fd94618fb454db4b9c46ec9eef1a5